+++ Sources/Sentry/include/SentryDispatchQueueWrapper.h
@@ -9,6 +9,8 @@ NS_ASSUME_NONNULL_BEGIN
 
 - (void)dispatchAsyncWithBlock:(void (^)(void))block;
 
+- (void)dispatchOnce:(dispatch_once_t *)predicate block:(void (^)(void))block;
+
 @end
 
 NS_ASSUME_NONNULL_END