+++ Tests/SentryTests/Networking/TestSentryDispatchQueueWrapper.swift
@@ -4,4 +4,8 @@ class TestSentryDispatchQueueWrapper: SentryDispatchQueueWrapper {
     override func dispatchAsync(_ block: @escaping () -> Void) {
         block()
     }
+    
+    override func dispatchOnce(_ predicate: UnsafeMutablePointer<Int>, block: @escaping () -> Void) {
+        block()
+    }
 }